trader actions updated
